+++ android/src/main/java/com/applovin/reactnative/AppLovinMAXNativeAdView.java
@@ -390,6 +390,7 @@ public class AppLovinMAXNativeAdView
         nativeAdInfo.putString( "advertiser", ad.getAdvertiser() );
         nativeAdInfo.putString( "body", ad.getBody() );
         nativeAdInfo.putString( "callToAction", ad.getCallToAction() );
+        nativeAdInfo.putDouble( "starRating", ad.getStarRating() );
 
         float aspectRatio = ad.getMediaContentAspectRatio();
         if ( !Float.isNaN( aspectRatio ) )
@@ -423,6 +424,7 @@ public class AppLovinMAXNativeAdView
         jsNativeAd.putString( "advertiser", ad.getAdvertiser() );
         jsNativeAd.putString( "body", ad.getBody() );
         jsNativeAd.putString( "callToAction", ad.getCallToAction() );
+        jsNativeAd.putDouble( "starRating", ad.getStarRating() );
 
         MaxNativeAdImage icon = ad.getIcon();
         if ( icon != null )